+++ lib/spicerack/input_object_base.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 module Spicerack
-  class InputObjectBase < Spicerack::AscriptorBase
+  class InputObjectBase < Spicerack::AttributeObject
   end
 end