+++ lib/spicerack/instructor_base.rb
@@ -5,7 +5,7 @@ require_relative "instructor/options"
 require_relative "instructor/core"
 
 module Spicerack
-  class InstructorBase < Spicerack::AscriptorBase
+  class InstructorBase < Spicerack::AttributeObject
     include ActiveModel::Model
     include ActiveModel::Validations::Callbacks
     include Instructor::Core