+++ CHANGELOG.md
@@ -8,9 +8,9 @@
 
 - Consolidated `RootObject` and abandoned gem
 - Consolidated `Instructor` and abandoned gem
-- Consolidated `Ascriptor` and abandoned gem
-- Refactored `Instructor` to be based on `Ascriptor`
-- Move `Tablesalt::Dsl::Defaults` to `Spicerack::Ascriptor::Defaults`
+- Consolidated `Ascriptor` as `AttributeObject` and abandoned gem
+- Refactored `Instructor` to be based on `AttributeObject`
+- Move `Tablesalt::Dsl::Defaults` to `Spicerack::Objects::Defaults`
 
 ### Rspice - [see changes](rspice/CHANGELOG.md#v0130)
 ### Tablesalt - [see changes](tablesalt/CHANGELOG.md#v0130)